Initial Commit
The basics of the website are there I guess. There is still a lot todo
for the presentation next week. Primarily, the multiple pages function
needs to be implemented and also the flex box needs to be figured out.
I’m using CSSslidy to make the slide show thing at the beginning.